Added text input
I changed the aesthetic to be more rustic since this game takes place in
a steampunk era society. Additionally I imported a new font which is a
more mono and typewriter style. Finally added a blinker for when the
user types into the box. Finally the max amount of characters that a
user can type is up to 30 characters to prevent sneaky players from
breaking the game